update on johns feedback
