+++ include/main.h
@@ -1,0 +1,50 @@
+
+/** Simple program "XYZ" realized.  
+ * @file main.h
+ * 
+ * 
+ * @mainpage Sort and search algorithms 
+ *  
+ *  @section sec1 Introduction 
+ *  This program written in C implements a well known sort and search algorithms. 
+ * 
+ *  @section sec2 Parts 
+ *  The program can be splitted into the following parts.
+ * 
+ *  @subsection sec2_1 Menu
+ * 
+ *  @subsection sec2_2 bla  
+ *  @subsection sec2_3 bla
+ *   
+ *  @author Gelbmann_Kristo_Valle 
+ *  @date 12.1.2021 
+ *  
+ *  @version 0.9 
+ *  Runs.
+ * 
+ *  @version 0.2 
+ *  Deleted dead code. 
+ * 
+ *  @version 0.1 
+ *  Added comments (Doxygen). 
+ * 
+ * 
+ * 
+ */
+
+
+#ifndef MAIN_H
+#define MAIN_H
+
+#include <stdio.h>
+#include <stdlib.h> // needed for function exit, dynamic memory mgmt
+#include <ctype.h> // needed for function tolower
+
+
+/** example to show how to do simple debugging
+ * 
+ * the way to comment here also here a test for doxygen - very very first draft
+   */
+#define DEBUG 1 //DEBUG_LEVEL
+
+#endif
